requirements.txt
flask
python-docx
xlsxwriter
requests

